Replaced error with NULL
the of_landings function can be a bit annoying since it current returns a stop error for data not found. This breaks a lapply call when working through a list of species or countries. The current workaround is the write a plyr::failwith to return a NULL. Making the function itself return a NULL eliminates this problem and makes life easier for a user. No need to write an additional wrapper for the landings call. So far it seems fine. Doesn't make sense to push to CRAN right away for such a minor bug fix so I'll see if I can add some additional functionality before doing so.